Add files via upload
1. What is React Native?
React Native is an open-source framework developed by Facebook for building mobile applications using JavaScript and React. It allows developers to build native mobile apps for iOS and Android platforms using the same codebase. React Native uses a declarative syntax and component-based architecture, making it similar to React for the web.

2. Key Concepts in React Native:
a. Components:
React Native applications are built using components, which are self-contained and reusable blocks of UI. Components can be either functional or class-based.
b. JSX (JavaScript XML):
JSX is a syntax extension for JavaScript recommended by React. It allows you to write HTML-like code in your JavaScript files, making the code more readable and expressive.
c. State:
State represents the internal data of a component. When the state of a component changes, React Native automatically re-renders the component.
d. Props:
Props (short for properties) are used to pass data from a parent component to a child component. Props are immutable and are used to make components dynamic and reusable.
e. Virtual DOM:
React Native uses a virtual DOM to efficiently update the UI. Instead of updating the entire UI when a change occurs, React Native updates a virtual representation of the UI and then updates only the necessary components in the actual UI.